[CDF-21820] 🧹Cleanup tests_migrations (#673)
* refactor; remove unused

* refactor: renaming to module upgrade

* refactor; fix

* docs; updated docs

* style: typo

* refactor: upgrade contribution guideline

* refactor: fix a few bugs in the run check script

* refactor: delete .venv after propject init is created
